Move SelectionKey into the selection map module
